Move ComplexDiff implementation to separate files (#1602)
